Add numpy packages.
While the issue seems to have been forgotten about, it was determined
that numpy available in bullseye wasn't usable. Instead, a dependency
was declared towards the 1.22.0 release.

The following versions are available in the Debian archive right now:

    numpy      | 1:1.19.5-1    | stable
    numpy      | 1:1.21.5-1    | testing
    numpy      | 1:1.21.5-1    | unstable
    numpy      | 1:1.22.1-1    | experimental

Backporting the package from experimental would require some heavy work,
so a “cherry-pick” approach was followed:
 - fetch the 1:1.22.1-1 package from experimental;
 - unpack its contents and its control files;
 - remove everything related to python3.10 from contents;
 - adjust Depends and Provides fields in control;
 - amend Version with a PiRogue identifier in control;
 - repack.