Adjust hostname for the PPA.
The piroguetoolsuite.github.io hostname is an implementation detail,
which /could/ change over time. In any case, this shouldn't be a
distraction for end users, who should find everything they need on
https://pts-project.org/ (documentation and PPA).

This commit doesn't add the new, intermediary debian-11/ directory
just yet (#3). Since the APT configuration is user-visible, it might
make sense to have some documentation about this change (needed to
prepare the migration to a Debian 12 base) published before doing so.