Revert "Move pirogue* directories under debian-11/ (#3)."
This reverts commit 500e0a0236bdb7feac6f9d09f8a55431237db123.

After assessing the situation further, it was decided to use the
following scheme for further releases:
  https://pts-project.org/debian-N/

They will be deployed via GitHub Pages from the debian-N repositories.

The existing PPA (based on Debian 11) will stay published at:
  https://pts-project.org/ppa/

and will also be published via an extra debian-11 repository at the
following location for consistency with further releases:
  https://pts-project.org/debian-11/

The ppa and debian-11 repositories will have to be kept in sync.